requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pydantic==2.5.0
httpx==0.25.2
Pillow==10.1.0
python-multipart==0.0.6
